ingest_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
